Спринт 15

+++ src/main/java/ru/practicum/shareit/user/UserService.java
@@ -7,11 +7,11 @@ import java.util.Collection;
 public interface UserService {
     UserDto createUser(UserDto userDto);
 
-    UserDto updateUser(Integer id, UserDto userDto);
+    UserDto updateUser(Long id, UserDto userDto);
 
-    void deleteUser(Integer id);
+    void deleteUser(Long id);
 
-    UserDto getUserById(Integer id);
+    UserDto getUserById(Long id);
 
     Collection<UserDto> getAllUsers();
 }
